Web Site
Update to Bootstrap 5, which I was already using for the Web REDUCE
GUI.  This has mainly required updates to the markup for accordions
and tabbed panels.  Bootstrap no longer requires jQuery, so now I only
load this on pages where I use it.  I have also updated the Web REDUCE
index page to follow the style of the rest of the site.  There are a
few minor visible differences from the previous version of the web
site, mostly caused by the design changes in Bootstrap.

Correct the way I use the PHP include statement to make it more
efficient by explicitly making filenames relative, thereby avoiding a
redundant search of the PHP include path.  Remove from the sitemap
JavaScript, Wasm and other Web REDUCE files that are loaded by the
index file.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5855 2bfe0521-f11c-4a00-b80e-6202646ff360