Web REDUCE: improved startup
Startup was unreliable due, I think, to a race condition between the
I/O Display iframe and JavaScript loading.  I now set the content of
the iframe in the JavaScript *after* setting up the handler for the
iframe load event, instead of in the HTML code.  This seems to be a
lot more reliable.  If that really proves to be the case then at some
later date I will remove the message about reloading the iframe.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6239 2bfe0521-f11c-4a00-b80e-6202646ff360