Web REDUCE: build mobile version
Add convenience targets webreduce and mini-webreduce to Makefile.  The
mini-webreduce target sets a smaller memory size and uses a version of
"package.map" in the same directory, based on the code in
"csl/new-embedded/minireduce-image".  The mini-webreduce build is
intended for mobile devices.

Update csl.cpp to use emscripten.h and EMSCRIPTEN_KEEPALIVE, but only
when building reduce.web.wasm, instead of specifying functions to
export in Makefile.  This is an experiment, based on some guidance I
read, which seems to work and might be better for future development.
Delete some redundant alternatively-named versions of the function
insert_buffer.  None of these changes should affect any builds other
than Web REDUCE.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6248 2bfe0521-f11c-4a00-b80e-6202646ff360