Correction and regression test file for Bug report: https://sourceforge.net/p/reduce-algebra/bugs/143/
With precise off, the input

 solve(1.2*(1.087^t)=9.3*(0.984^t),t);

causes an infinite recursion.
This happens because an expression of the form (a/b)^c is simplified to a^c/b^c when precise is off, while inside solve the expression

   4*1087**t - 31*984**t = 0

is replaced by

   4/31 - (984/1087)**t = 0

by a rule, and immediately turned back into the original form by simpexpt.

The immediate problem can be solved by setting !*precise to t when using solve rules. For the recursion I introduced a counter and a check that raises an error when this
counter exceeds a certain limit (default 20) of rule applications.



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6261 2bfe0521-f11c-4a00-b80e-6202646ff360