REDUCE on Common Lisp: revise the build process
Streamline the building of bootstrap REDUCE: replace most of the code
in "bootstrap.sh" with Lisp code in "bootstrap.lisp"; merge the code
in "build.red" into "bootstrap.lisp" after converting it to Standard
Lisp; merge the remaining code in "bootstrap.sh" into "build.sh". The
shell script "build.sh" now has a -b option that allows it to replace
"bootstrap.sh" and provides better help messages.

Abandon the attempt to replace "bootstrap.sh" entirely with Lisp code
because cannot comfortably compile and load in the same Lisp session
due to defconstant errors. But probably not feasible to build REDUCE
entirely with Lisp code anyway. Using a single shell script seems a
reasonable compromise.

Attempt to fix a package locking error that manifests itself on Gentoo
Linux, but I can't test this because it doesn't happen on the versions
of SBCL that I currently run.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6284 2bfe0521-f11c-4a00-b80e-6202646ff360