REDUCE on Common Lisp
Mainly improved support for CCL.  Revise the system function to try to
return the exit code, but it doesn't seem to work on Windows and I
have not yet tested it on Linux.  Correct the pwd function.  Revise
the cd function.

Automatically accommodate either command name ccl or ccl64 for both
building and running.  Remove a spurious director separator in
"red*.bat".


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6295 2bfe0521-f11c-4a00-b80e-6202646ff360