Minor update to pm package
Rainer's previous update to pmintrfc.red to avoid redundant warnings
that rset and rsetd are already defined as algebraic procedures
involved calling infix in symbolic rather than algebraic mode.
However, this prevents the source files being input with "on defn"
because the affected infix calls are no longer evaluated when they are
input, and this breaks the Common Lisp build, which uses "on defn" to
generate Lisp code to be compiled.  So I have reverted that change and
instead suppressed the warnings by turning msg off and on again using
assignments, which generate Lisp with "on defn".  This fix seems to
work correctly for REDUCE on CCL, PSL and CSL.

I have also removed some redundant code from the end of pm.red and a
redundant procedure parameter in pattperm.red.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6297 2bfe0521-f11c-4a00-b80e-6202646ff360