REDUCE Manual
Revise camal.tex to work with MathJax: comment out use of the
subequations environment, which is not supported by MathJax and in the
PDF version only affects the way the equations are labelled; replace
\ref with \eqnref, which is defined in manual.tex.  Also, remove the
"h" at the start of the file that clearly should not be there!  Fix
some inconsistent horizontal spacing and a typo in appendd.tex.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6346 2bfe0521-f11c-4a00-b80e-6202646ff360