Web Site: replace RSS feeds
SourceForge RSS feeds are no longer accessible by PHP code running on
the web server, so I can no longer display them on the home page.
This is why the RSS feed displays have been frozen for the last few
months; they were being served from the cache which was not being
updated.  I have therefore replaced the RSS feed displays with a list
of links to useful SourceForge pages, two of which provide the same
information that was previously displayed via the feeds.  This allows
me to remove a lot of PHP code that is no longer used.  I have also
updated Bootstrap to version 5.2, which is the latest stable version.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6354 2bfe0521-f11c-4a00-b80e-6202646ff360