Interactive Lessons: fix a syntax error
The addition of a Subversion identifier line (containing $Id: ... $)
to the top of each of the interactive lesson files caused a syntax
error because the first $ terminated the comment statement that the
line was in and then REDUCE tried and failed to parse the rest of the
line.  I have added a ; at the start of each Subversion identifier
line, which terminates the comment statement and leaves the rest of
the line in a % comment, as at first glance it appeared to be but
wasn't!


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6404 2bfe0521-f11c-4a00-b80e-6202646ff360