Definite integral of sqrt
The definite integral int(sqrt(x*(1-x)),x,0,1) evaluates to -pi/8 but
is obviously positive!  The substitution x = (1 + sin theta)/2
facilitates evaluation by hand to +pi/8.  REDUCE evaluates this
integral by indefinite integration to a complex-valued expression
involving logarithms, followed by substitution of the limits.  Doing
this interactively gives the result (-i log i)/4.  Noting that log is
multi-valued in the complex plane, the result can be expressed as pi/8
+ n pi/2 for any integer n, and choosing the principal branch gives
the correct result pi/8.

However, when REDUCE performs this calculation automatically, it calls
mkdint in "int/dint.red", which sets "off precise" to avoid ugly
results involving abs.  Eventually, defint11s in "defint/defintx.red"
is called, which splits out any factor of the integrand that is
independent of the integration variable.  For the integrand in
question, with "off precise", it splits out a factor of the imaginary
unit and changes the form of the integrand into a product of square
roots.  I think it is this change to the integrand that leads to the
integral having the wrong sign.

One solution would be to set "on precise": doing so and then calling
defint0 by hand gives the correct result.  However, this is likely to
lead to ugly results involving abs in similar but more complicated
cases.  So the solution I have chosen is to test whether the factor of
the integrand that is independent of the integration variable is pure
imaginary, and if so ignore the factorization and pass the whole
integrand to the integrator via defint11.

I have also modified defint11 slighlty to avoid unnecessary
simplification of the integrand before passing it to simpint,
otherwise the fix to defint11s doesn't work.

The revised code runs the test file exactly as before, but also
evaluates the integral in question correctly, which I have added to
the test and log files.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6492 2bfe0521-f11c-4a00-b80e-6202646ff360