Minor manual corrections
Fix glitches in a couple of updates to the REDUCE manual made over the
last year or so, one in oper.tex and one in solve.tex.  These were
spotted by a vigilant reader of the HTML version on the web site and
reported to me.  (People do read manuals!)  I'll also update the HTML
and PDF versions of the manual on the web ASAP.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6554 2bfe0521-f11c-4a00-b80e-6202646ff360