A month or so ago I checked in updated to the CSL configuration files that were to provide better support for cross compilation. At the same time they tidied away some features there were obsolete. It turned out that some of the changes hurt builds on Windows (in ways that I failed to notice until very recently). This should put things back in better order.
I have also introduced a new notation: where previously one would write
"lambda x; A" one can now use "[x] A", a syntax more in line with some
other current languages. The oddity here is that "[" is to a large extent
treated as a synonym for "<<" so in general use "[a,b,c]" and "[a;b;c]"
both parse into use of the Lisp-level group operator PROGN. So up until
now "[x] A" will have led to a diagnostic "progn x invalid as operator". The
new overload that interprets it as a lambda expression (ie anonymous
function) is only used if all the items with the square brackers are
symbols.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6559 2bfe0521-f11c-4a00-b80e-6202646ff360