I try to improve Unicode support in CSL here. This checkin is tested on Linux only - and once it is in place I will see how Windows and Macintosh go, since they may be broken by these changes.
The adjustments are intended to be
    charnameALT-x   replaces the item with the (typically unicode)
                    character named. This has been available with the
                    console mode version of CSL for some while, but
                    I hope it now works in the GUI.
    #charname;      is interpreted as the character so named, with the
                    conversion performed within the Reduce parser so it
                    is less CSL-specific.
The two above schemes provide alternative ways of getting extended characters
into the system. I think maybe the first is nicer for interactive use while
the second can go in files so that the files are remain in a base character
set.
    The tables in tmprint.red have bee updated to try to make the support
    for Greek letters a bit nicer.
    "on/off raise and lower" in CSL will now only impact Basic Latin
    chararacters. Previously they would SOMETIMES and depending on the
    platform change the case of Greek letters etc. The main way if which
    this may cause grief is that they will not now inpact the case of
    accented letters (eg "e-grave" and so on). If anybody finds that
    hurtful please contact me to explain what will work best!
There are around 2000 names for "charname" that are recognised, and they
should be in agreement with HTML conventions. I have only done a little
testing and mainly with alpha, beta, gamma, Alpha, Beta, Gamma etc.
OK - now I will check on Windows and Mac.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6678 2bfe0521-f11c-4a00-b80e-6202646ff360