Minor manual additions
Rename the SHOWTIME Command section to Timing Facilities, include the
time switch here, and add the use of time().  Tidy the File Handling
Commands chapter and add a section on Using Variables as Filenames.
Add a remark about the effect of off nat to the end of the section on
write.  Update the PDF and HTML manuals on the web.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6691 2bfe0521-f11c-4a00-b80e-6202646ff360