Using ln to represent natural logarithm (again)
Change the manual to say that REDUCE defines the identifier ln to be
an abstract operator with no properties.  This seems safer than
removing that definition, since it is used in a couple of test files,
so users might expect it.  Also, adjust unzip-manual.sh to be
executable from any directory.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6700 2bfe0521-f11c-4a00-b80e-6202646ff360