REDUCE IDE version 1.11
Add a new option `reduce-run-terminal` to specify the value of TERM to
use on Unix-like platforms so that CSL REDUCE responds appropriately
to interrupts, which with a dumb terminal it does not.  It defaults to
nil on MS Windows and to `Eterm` on all other platforms.  There are
also quite a lot of other minor updates; see the end of "README.md".


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6711 2bfe0521-f11c-4a00-b80e-6202646ff360