Working directory support
Revise redcsl.bat so as not to pass the command argument -nocd on to
CSL, which complains about it.  Add a new web page at
https://reduce-algebra.sourceforge.io/directories.php describing
working directories in REDUCE.  Corrections and other comments on this
page are welcome!

Also set the svn:ignore property on web/htdocs/reduce-ide to ignore
some non-source files.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6716 2bfe0521-f11c-4a00-b80e-6202646ff360