REDUCE IDE Version 1.12
Fix a bug in the display of the current procedure name.  Improve
reduce-tagify-dir-recursively, which previously failed on the REDUCE
packages directory.  Tidy up the way REDUCE IDE runs REDUCE on
Microsoft Windows and retain the current working directory in CSL
REDUCE.  General improvements to the way REDUCE IDE runs REDUCE that
allow it to be run directly without involving a shell, which is most
useful on Microsoft Windows.  Update the web site.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6734 2bfe0521-f11c-4a00-b80e-6202646ff360