The input
--b;

leads to an error message about car of nil:

+++ Error attempt to take car of an atom: nil

This happens because we have now an operator "-->" defined in the core system, but no operator definition for "--".
Previously the input --b parsed as (minus (minus b)).

There is at least another case, namely ".~*", so that ".~b" gives the same error.

This commit triggers a Reduce error message instead of trying to take car of nil. It is not a complete solution for the problem.
I think the parser needs to backtrack when it encounters this situation.



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6748 2bfe0521-f11c-4a00-b80e-6202646ff360