Web REDUCE Gnuplot support (part 2)
Add some debugging support for gnuplot: in debug mode it is possible
to send input directly to gnuplot. Revise the Web REDUCE typescript
files to use strict type checking and generally make the code more
robust. Tweak the REDUCE gnuplot package (currently by redefining some
code after loading gnuplot) so that the gnuplot program is run via the
"system" function and not via a pipe. The underlying emscripten
version of std::system normally returns without doing anything, but I
have added code (see the previous commit) to detect a request to run
gnuplot, which it does by sending a message from the worker thread
back to the main thread.

The reason for two separate commits is that I develop the main web
framework on Windows but for Unix-oriented build processes, as used by
both REDUCE and Gnuplot, I find it easier and faster to run emscripten
on Linux.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6784 2bfe0521-f11c-4a00-b80e-6202646ff360