New package Logo Turtle Graphics
LogoTurtle is an alternative to the Turtle package.  It aims to be
more efficient, more authentic, potentially more interactive and
potentially more complete than Turtle.  It currently provides all of
the facilities provided by Turtle but I plan to develop it further.
It should work with all versions of REDUCE that support Gnuplot,
including Web REDUCE (although Web REDUCE does not currently support
pause, making it awkward to run the test file).

I have added a section to the REDUCE manual, and also indexed Gnuplot
and Turtle as packages.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6802 2bfe0521-f11c-4a00-b80e-6202646ff360