A small step forward in the sandbox code that probably shows anybody who reads it more aboy where it os going but that is still well short of making it work. And the previkous version using fork() has been removed (anybody who wants that can recover an earlier revision).
One of the pending nits of tidying up is that with the current
conservative GC the class Save and its method restore() become
noops and I will be able to remove all instanced of them. Well there
are perhaps 830 of them spread almost everywhere so that will take
a while. but I believe it will make a fair amount of code cleaner.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6817 2bfe0521-f11c-4a00-b80e-6202646ff360