Update Logo Turtle Graphics package
This is an incompatible user interface change.  Move the facility for
DRAW to accept optional arguments into LOADPICT.  Move the facility
for WRAP, WINDOW and FENCE to accept optional arguments to new
commands SETWINDOWSIZE and SETTURTLEMODE, and introduce matching
queries WINDOWSIZE and TURTLEMODE.  Commands are now less overloaded
and none take zero or more arguments, although some still take one or
more.  Make the commands that never take any arguments into endstats,
so that they are not required (or allowed) to be followed by empty
parentheses.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6850 2bfe0521-f11c-4a00-b80e-6202646ff360