This checkin fixes two subtle bugs in floating point/integer conversion.
The first bug arises from the different precision of integer and floating point arithmetic.
When a floating point number is converted to an integer, a check is done whether the resulting integer fits in the fixnum range, i.e., into a machine word. 
This is done by comparing it to the largest and smallest floating point numbers that can be converted into a fixnum. If the fixnum word size is n bits, the largest
integer is 2^(n-1)-1, the smallest integer is -2^(n-1). The bug arises from the fact that an IEEE double float has 53 bits of mantissa length, so that 2^(n-1)-1 cannot be represented
if (n-1) > 53. In this case the largest floating point number that can be converted into an integer is 2^(n-1) - 2^(n-54), a positive integer that has exactly 53 1-bits.
This is the case on 64 bit architectures, like x86_64 or aarch64.  All fixnums larger than this are represented by a floating point number with 53 significant digits,
with resulting loss of precision.  But when converting such a larger float to an integer, the result doesn't fit into a fixnum and must be converted into a bignum.
The bug occured in such a situation: when converting a float into an integer, the float is compared to the result of the largest floating point number that can be converted into a fixnum.
In 64bit architectures, this floating point number was computed incorrectly as (2^63-1) instead of (2^63-2^54) to decide whether to return a fixnum or a bignum. For floats between these
two numbers the conversion to a fixnum failed. On x86_64, the conversion operation returns -2^(n-1), ie. a 1-bit followed by 63 0-bits (this is what Intel called indefinite integer value).
I.e., converting a large positive float to integer returned a large negative integer.

The second bug is an implementation detail: after the bignum code is loaded, calling fix on a fixnum returned a bignum if the fixnum did not fit into a halfword.
THis was done for optimization of certain operations that can be performed without loss of precision or overflow. A simple redefinition of fix solves this.
I made the change only for 64 bit systems; I still have to check the 32bit code.



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6885 2bfe0521-f11c-4a00-b80e-6202646ff360