I hope that alg/simp.red now counts the number of items tagged as negative in arguments to abs, so that "let sign(a)=>-1,sign(b)=>-1;abs(a*b*c);" behaves better. Others may want to check both behaviour and the code change I put in please.
rtools/general.red now provides a function "gensym1" that can be given
an argument and returns an uninterned symbol based on that but with a
sequence number in the stle of regular gensyum().

cslbase/bytes2.cpp addresses just how catch and throw behave in CSL
bytecoded functions and I believe it fixes a bug there.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6888 2bfe0521-f11c-4a00-b80e-6202646ff360