Minor manual revisions
Try to use => instead of = consistently in let rules, etc.

Move the listproc example to the section on the LISTVECOP package and
replace it with a simpler example that does not use the LISTVECOP
package.  Add a remark that _ is a postfix operator and that it must
be separated from the identifier that it follows.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6893 2bfe0521-f11c-4a00-b80e-6202646ff360