New package WITH to support local switch settings
This package defines an infix operator `with' that implements the
syntax

   expression with on/off switches, on/off switches, ...

where `on/off' is on or off and `switches' is a single switch name or
a comma-separated sequence of switch names (as for the on and off
commands).  The `with' operator returns the expression specified as
its left operand evaluated and displayed subject to the switch
settings specified as its right operand, applied locally.  For example:

   pi with on rounded;
   (a+b)^2/2 with off exp, on div;

The package is a single file in the "misc" directory that is
documented as a user contributed package in the manual.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6898 2bfe0521-f11c-4a00-b80e-6202646ff360