Prevent my new WITH operator from autoloading
It seems that WITH upsets EXCALC so I have commented out the autoload
code in "entry.red" while I investigate.  To use the WITH operator
without the autoload code it is necessary to execute the code that I
commented out, which is also included near the top of "misc/with.red",
and then do load_package with.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6899 2bfe0521-f11c-4a00-b80e-6202646ff360