Minor manual and web updates
Fix documentation for "with" package to look better on the web.

Change line breaking in titles of some user-contributed packages so
that they are not broken in the TOC, and change $n$ to \emph{n} in the
title of the Lie package.  This makes the list of user-contributed
packages at https://reduce-algebra.sourceforge.io/documentation.php
look much better.  (The PDF bookmarks still look correct.)

Split web/copy-manual.sh into two files, copy-manual.sh and
upload-manual.sh, to make it more flexible, and explain how to use
these scripts in the README.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6927 2bfe0521-f11c-4a00-b80e-6202646ff360