Web Site: add "missing manuals"
Add to the web site PDF versions of the documentation listed at the
top of `doc/manual/user.tex` as "missing" provided the packages they
relate to are pre-compiled in distributed versions of REDUCE.  This
commit includes only source files and only the PDF files are included
on the web.  Later, I will link to these PDF files from the manual
stubs where appropriate.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6931 2bfe0521-f11c-4a00-b80e-6202646ff360