Here I have some changes that are a framework for a project that might be tolerably limited in scope but that it would be REALLY good if somebody else picked up and ran with it.
The overall objective is to take existing Reduce/RLisp source files and
adjust line breaks and indentation to standardise them and make them
prettier and possibly more uniform from package to package.

So this is what I have done, and what needs doing to follow through.
(1) rlisp/tok.red now arranges that when a token is read (typically by
a call to scan()) the variable comment!* gets set to a list of strings that
are any comments seen while reading it. Well this happens if !*rprint is
true, so in the default case when that is nil there is nothing extra going on.
It is at least possible that what happens here is about all that will be
needed - but when everything else is complete the details here will
deserve review. A clear issue is that of what happens about material
skipped because of "#if" directives.

(2) in rlisp/xread.red, xread() behaves in the old way again unless !*rprint
is set. If that is set I want it to end up such that if the input is
say "(x+b) % comment" that the parse tree returned picks up the value in
comment!* and hands back (!% (plus x b) "comment"). At present the places
where comment!* is checked are very very limited and so comments only get
attached in a very few places and not obviously where they really should.
There are two bits of re-work called for - one is to detect comment!*
everywhere and about as soon as it can be, the other is to think about the
comments in "f/*c1*/(/*c2*/a/*c3*/,b)" and just how the multiple comments
can end up placed in the Lisp form of that function call. In particular while
c1 and c3 readily attach to f and a, c2 does not and so recording it (and
any subsequent ones after the ",") needs some new notation. Possibly
"(!%!% c2)" inserted where it looks like a first argument for the function
call but later on will be detected as something special. It will be important
that xread (and friends such as the code that parses "for" loops behave just
as they currently do unless !*rprint is set.

(3) The output from (2) is not always going to be directly executable Lisp,
but it is not intended for execution, only for printing. In rlisp/superv.red,
rlisp/proc.red and rlisp/form.red I disable various conversions and expansions
that are normally done so as to preserve the structure of the original input.
And what is left gets passed to the code in rprint/rprint.red that displays
it as Reduce source code. This code needs to be extended so that it detects
and handles "%" forms (and possibly "%%" ones so that comments are placed
in the generated nice version of the code. After doing that it may be that
there are some general improvements to the layout to consider, but it is
at least not a disaster at present. Forms with "%" in should only arise
when (1) amd (2) have introduced them.

Parts (2) and (3) could be worked upon almost independently. xread.red is
only 312 lines long so even though it is a bit ugly (it may have to be
coded in a subset of rlisp because it is processed early in a bootstrap build process) it should not be too hard to get to grips with. rprint.red is longer
at 853 lines but the vast majority of those are just handling various
special cases and so it splits into a large number of independent paragraphs.

So will somebody please have a go at this - and email in to say that they
some sample adjusted versions of those files....  



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6937 2bfe0521-f11c-4a00-b80e-6202646ff360