Manual: minor formatting revisions
These revisions are primarily for the HTML version.  Remove file
bchr8t.htf since the problem that it fixed has been fixed at source.
Define \texttt to generate <code></code> when using TeX4ht so as to
avoid the space after it being set in a monospace font and so
elongated.  Replace \terminalfont by \texttt, so that commands that
used \terminalfont also avoid elongated spaces after them.  Revise
some choices of font and spacing in Chapters 1 and 2 (which apply also
to the PDF version).


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6952 2bfe0521-f11c-4a00-b80e-6202646ff360