Manual: Entries for MATHML and MATHMLOM packages
Expand the stubs into complete entries that no longer link to separate
PDF files.  Add package "with" to "Changes since Version 3.8".  Add
the fix to avoid long (monospaced) spaces after \texttt{} when using
TeX4ht to the configuration file used to generate the HTML version of
the REDUCE manual that is included with snapshot releases.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6955 2bfe0521-f11c-4a00-b80e-6202646ff360