Manual: Add SVN revision number
Add the revision number just above the date on the title page.  The
revision number is found automatically in a new bash script called
"revision.sh".  This runs svnversion, reprocesses its output and
writes it to a temporary file called "revision.tex", which is deleted
at the end of the run.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6959 2bfe0521-f11c-4a00-b80e-6202646ff360