LINALG package: simplex
By default, simplex throws an error if a problem has no feasible
solution.  (It is debatable whether no solution is actually an error.)
Introduce a new switch noerrsimplex such that when on
(it is off by default) simplex returns the empty list {}, which
facilitates use of simplex as a subroutine within other code.  Tidy up
a few issues with the documentation and add this new switch.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6970 2bfe0521-f11c-4a00-b80e-6202646ff360