Weierstrass p-functions
Trying to plot weierstrass1(u, 1, 0) fails.  One reason is the
Catch-22 situation that weierstrass1 only evaluates numerically with
on complex, but plot turns complex off.  The values of weierstrass1(u,
1, 0) for real u are real, and it seems to me to be inappropriate to
need to turn on complex explicitly to evaluate a real-valued function
of a real variable.  Therefore, I have modified the code to turn
complex on internally when rounded is on.  I regard this very much as
an experiment.  There may well be a better solution.  If it is the
right change then it should be applied more generally to the other
related functions.  I don't see any change to the output when I run
the test file.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6976 2bfe0521-f11c-4a00-b80e-6202646ff360