Weierstrass p-functions
Fix a minor over-enthusiasm in my previous commit.  With this fix,
running "ellipfn.tst" agrees with "ellipfn.rlg" except for the value
of e1^4+e3^4-e2^4-e4^4, which was 1.42679947108e-16 and is now
-3.63207759546e-15.  I'm hoping that it should be zero, so that the
new result is no more incorrect than the old result!  Further such
changes are likely if I manage to further optimize the numerical
evaluation.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6980 2bfe0521-f11c-4a00-b80e-6202646ff360