The numerical evaluation of the Weierstrassian and Jacobi elliptic functions required that both ROUNDED and COMPLEX were ON even if all arguments were real. As Francis Wright pointed out that this causes problems when attempting to plot these functions as gnuplot turns COMPEX OFF! In an earlier commit he made some changes to efweier.red and efnumeric.red to work around this difficulty when evaluating weierstrass1 and weierstrassZeta1.
In this commit similar changes are made to efjacobi.red to allow the numerical
evaluation of sn, cn, dn and friends when only ROUNDED is ON. This is
permissible as these functions are all realvalued.
Also the functions weierstrass1 and weierstrassZeta1. are flagged as realvalued
in efweier.red.crack/applysym.red

The manual pages  ellipfn.tex have been updated to document these changes.

It may be possible to extend these changes to include some elliptic integrals
and inverse Jacobi functions where these are realvalued or realvalued subject
to range constraints on the arguments. This is currently being investigated.

Currently for some real-valued functions  the numerical algorithms occasionally
return a spurious imaginary part due to rounding. This is always far smaller
than the current precision.

In a future commit is planned to remove this spurious part by post-processing
in the function n_numeric. Similarly in other cases where it is known that
the result should be purely imaginary, it should be possible to remove any
spurious real part.




git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6981 2bfe0521-f11c-4a00-b80e-6202646ff360