GNUPLOT package
Replace the somewhat arbitrary sets of unset commands issued before
each new plot with the reset command, which is much tidier and makes
it easier to add new options.

Add examples of using logscale and pm3d, and an example of a temporary
hack to plot a function with a singularity.

Fix use of procedure ikeda to avoid
***** Binding stack overflow, restarting...
and label the vertical axis sensibly.

Update the manual to make the relationship between the REDUCE GNUPLOT
package and the gnuplot graphics utility clearer; tidy up the
typography a bit; and document the logscale, pm3d and style options.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7008 2bfe0521-f11c-4a00-b80e-6202646ff360