GNUPLOT
Move yesterday's revision to procedure ploteval2xyimpl one level up
the call hierarchy into procedure ploteval3xy, which is more elegant
and includes 3D surface plots.  So it is now possible to avoid the 3D
plots with left-handed coordinate systems that appear in some of the
example plots in the manual, as well as the 2D contour plot with y
horizontally and x vertically.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7029 2bfe0521-f11c-4a00-b80e-6202646ff360