GNUPLOT
Use z-range to flatten 3D plots at the limits, which is particularly
useful for functions with singularities.  This avoids a surface plot
appearing as a stack of 2D plots in some cases, which I think is a bug
that this change essentially works around, but which I need to
investigate further.

Document this change, the use of plot_xrange etc., and the fact that
refine applies only to 2D plots.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7033 2bfe0521-f11c-4a00-b80e-6202646ff360