ellipfn/efjacobi.red:  Corrections to the numerical evaluation algorithms which use real arithmetic (when both arguments are real and k <1).  The original algorithms for dn, nd, sd, ds, cd and dc experienced large rounding errors when the first argument of these functions was near to K(k) and the implementation  of the procedure otherdn designed to overcome these problems was found to be incorrect. This problem has been now been corrected.
Minor changes to the procedure tidy_result ellipfn/efnumeric.red to improve its
behaviour.

Changes to efellint.red  and efjacinv.red to use simpler faster algorithms
using only real arithmetic where possible; typically when k is real and |k|<1
and when there are range restrictions on the first argument of the incomplete
integrals and inverse functions.  Instances of calls to rederr have been
replaced  by lisp error so that these may be caught with errorset. This should
facilitate the graphing with gnuplot of functions with singularities.

Minor changes to ellipfn/ellpfn.tst and ellipfn.rlg.

Changes to the manual pages in ellipfn.tex to document recent changes: most
importantly that for numerical evaluation to occur it is only necessary that
the switch ROUNDED is ON and that the arguments of the elliptic functions are
numeric.  Of course, if any argument is complex, the switch COMPLEX must also be
ON, for the arguments to be recognised as numeric. In some cases even when
arguments are all real, the numerical evaluation algorithms require COMPLEX to
be ON (For example when |k|>1 or the first argument of arcsn >1).  In these
cases COMPLEX is turned on automatically for the duration of the evalaution and
then returned to its former value.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7037 2bfe0521-f11c-4a00-b80e-6202646ff360