ellipfn/efjacobi.red:  Generalisation of the numerical evaluation algorithms which use real arithmetic (now when both arguments are real).
Minor changes to the procedure n_elliptic in ellipfn/efnumeric.red to improve
its behaviour.

efweier.red: Introduce a new function weierstrass_sigma0 which is a version of
weierstrass_sigma with lattice invariants as the 2nd and 3rd
parameters (cf. weierstrass1 and weierstrassZeta1).
Add various rewrite rules for this function, declare it  real-valued and add a
fancy-prifn etc for this function.
Shorten the output of weierstrass_sigma1, 2 and 3 on plain terminals.
Correct the rules triggering the numerical evaluation of weierstrass,
weierstrassZeta and weierstrass_sigma and remove the erroneous declaration
of these as real-valued.

Minor changes in efjacinv.red to the algorithms that use only real arithmetic.  

Update to ellipfn.rlg.

Changes to the manual pages in ellipfn.tex to document recent changes: most
importantly that for numerical evaluation to occur it is only necessary that
the switch ROUNDED is ON and that the arguments of the elliptic functions are
numeric.  Of course, if any argument is complex, the switch COMPLEX must also be
ON, for the arguments to be recognised as numeric. In some cases even when
arguments are all real, the numerical evaluation algorithms require COMPLEX to
be ON (For example when |k|>1 or the first argument of arcsn >1).  In these
cases COMPLEX is turned on automatically for the duration of the evalaution and
then returned to its former value.
Correct numerous typos in the text; in particular change eta_1 in the manual to
eta1 (as in the code). Update and correct the formatting of a couple of tables.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7039 2bfe0521-f11c-4a00-b80e-6202646ff360