Plotting Weierstrass Elliptic Functions
Add ellipfn/efweierplot.tst, which tests both the Weierstrass elliptic
functions and the GNUPLOT package.  It aims to reproduce as closely as
possible the plots in the NIST Digital Library of Mathematical
Functions at <https://dlmf.nist.gov/23.4>.  Note that this file
includes a lot of Unicode characters in the comments, which I copied
from the web page.  The results are OK although some of the plots are
slow.  Also, this test shows up a problem with plot, which I will
investigate, although there is a simple workaround.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7042 2bfe0521-f11c-4a00-b80e-6202646ff360