ellipfn/eftheta.red:  Generalise the numerical evaluation algorithms of the four theta functions and their derivatives so the second parameter is either interpreted as 	    the modulus (tau) when Im(tau)>0, 	    the nome (q) when q is real and lies between 0 and 1. For other values of this parameter the functions are not defined and the numerical evaluation fails.
The conditions in the rules triggering numerical evaluation have been changed:
the switch ROUNDED needs to be ON when both parameters are recognised as numeric.
Thus if the argument and the nome are both real, only ROUNDED needs to be ON.
If the second argument is to be interpreted as the modulus the switch COMPLEX
must also be ON and similarly for a complex argument and a real nome. These changes
should facilitate the evaluation and graphing of theta functions when both the
argument and nome are real, which is the case in most practical applications.

Added two new files: efjacobiplot.tst and efthetaplot.tst to produce graphs similar
to those on https://dlmf.nist.gov/22.3 and https://dlmf.nist.gov/20.3 respectively.

efweier.red:  add calls to the procedure tidy_result to remove spurious
(insignificantly small) real or imaginary parts from the results returned by the
lattice-generator and lattice-invariant functions etc. 

Changes to the manual pages in ellipfn.tex to document the recent changes to the
behaviour of the theta functions and their derivatives and to correct minor typos.



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7058 2bfe0521-f11c-4a00-b80e-6202646ff360