Unset command
This does not work as described in the manual.  So I have replaced it
with a version that I claim does.  My replacement is rather
simple-minded and I may have overlooked some situations where the old
version is required, so I have left the old code in the file but
commented it out.  I have also tweaked the manual to note that unset
is like clear and include links to clear and mkid.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7078 2bfe0521-f11c-4a00-b80e-6202646ff360