Unset command
Revise yesterday's commit.  Restore the previous version of unset and
merge yesterday's revision into it.  Now unset works as before but
also works with mkid, as described in the manual.  I have also
documented the operation of the previous version, which indirectly
clears the value assigned to the value of an operator.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7079 2bfe0521-f11c-4a00-b80e-6202646ff360