Experimental new memory layout for PSL - work in progress
The current memory layout requires the memory region used for compiled lisp code to be both writable (for loading compiled code into memory) and executable.
This is not possible in some operating systems (OpenBSD and MacOS).  In both cases the solution must use mmap to allocate a memory area that can either be mapped twice (OpenBSD) or switched between writable and executable (MacOS).  In order to do this, compiled lisp code must be position independent and therefore not contain compiled static lisp expressions.
The latter will be allocated in a separate, writable, but not executable memory region.
I will add a detailed description of the old and new memory layout in the near future.

This checkin is for x86_64 only. It defines a new architecture AMD64_openbsd that implements the necessary changes to the PSL comüpiler in order to generate a working bpsl kernel for the new memory layout, successfully tested on OpenBSD.  I'm currently working on compiling into memory and generating fasl files.

Ultimately this will become the memory layout for all 64 bit architectures. I hope I can make it work on ARM-based Macs.



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7109 2bfe0521-f11c-4a00-b80e-6202646ff360